requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dss-python==0.10.7.post1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
future==0.18.2
ipykernel==6.9.1
ipython==8.1.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
llvmlite==0.38.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.6.3
numba==0.55.1
numpy==1.21.5
OpenDSSDirect.py==0.6.1
packaging==21.3
pandapower==2.8.0
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
scipy==1.8.0
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
